Show Alert Dialog for deleting and canceling the alert dialog

+++ app/src/main/java/com/omega/Fragments/CheckoutFlashCardFragment.java
@@ -1,5 +1,6 @@
 package com.omega.Fragments;
 
+import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
@@ -89,9 +90,26 @@ public class CheckoutFlashCardFragment extends Fragment {
 
         @Override
         public void deleteItem(int adapterPosition) {
-            Groups group = groupsAdaptor.deleteRowFromDataSet(adapterPosition);
-            flashCardViewModel.deleteGroup(group);
-            Log.d(TAG, "deleteItem: Row Deleted");
+
+            Groups group = groupsAdaptor.getItemAtPosition(adapterPosition);
+
+            AlertDialog.Builder alertDialog = new AlertDialog.Builder(getActivity());
+            alertDialog.setTitle("Confirmation Alert");
+            alertDialog.setMessage("Are you sure you want to delete item ? ");
+
+
+            alertDialog.setPositiveButton("Yes", (dialog, which) -> {
+                flashCardViewModel.deleteGroup(group);
+            });
+            alertDialog.setNegativeButton("Undo", (dialog, which) -> {
+                groupsAdaptor.refresh(adapterPosition);
+            });
+
+            alertDialog.setOnDismissListener(dialog -> groupsAdaptor.refresh(adapterPosition));
+
+            AlertDialog dialog = alertDialog.create();
+            dialog.getWindow().setBackgroundDrawableResource(R.color.DarkModePrimaryDarkColor);
+            dialog.show();
         }
 
         @Override
